+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r105-beta...f448243d9130c45dba36382f9273613500f59c64)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r106-beta...7460568a8d93c5ae3cf076de821c862b8014d4a6)
 
-- Fix a lib init error
-- Cell不是CompactRaid修改版或增强版！
+- Update M+ season 4 debuffs
+- Add Current Season dropdown item
+- Update RaidDebuffs
+- [snippet] MythicPlusDebuffs
+- Update About